langgraph/requirements.txt
langchain_core
langgraph 
langchain_community 
tavily-python 
python-dotenv 
langchain-groq
IPython